Create ConversationManager and discover underlying objects
ConversationManager is responsible for the delivery of a greeting. It
instruments an ITokenReader wrapped inside an IExceptionHandler to
protect against unforseen exceptions. Exceptions trapped in the
IExceptionHandler will be logged by an ILogger. For UnitTesting
purposes, the ExceptionHandler class needed to be implemented, as it
would otherwise not be possible to properly unit-test the logic