ExceptionHandler runs Func<T> and logs exceptions
ExceptionHandler is a class that is designed to run functions
(Func<T>) and Actions within a run method. It executes in a try/catch
statement and will use ILogger to log what happened.
Implementing the ExceptionHandler forced a change to the
ConversationManagerTests, where one of the UnitTests was uncalled for
in the scope of ConversationManager. Specifically, verifying that
exceptions are logged is the responsability of the ExceptionHandler and
not the ConversationManager.